lw: node labels and attached labels now scale somewhat correctly with the view.

+++ src/clients/legacyWatcher/manetglview.h
@@ -140,7 +140,7 @@ signals:
 
         struct DisplayStatus
         {
-            float scaleText;
+            int scaleText;
             float scaleLine;
             int monochromeMode;
             int threeDView;
